Quickfix to make microfactories playable
This is just a stupid quickfix to make the top left corner
of the inventory of the player's hold visible in the forms of
the various microfactories.

A proper fix would be to generate and set the form when the
user uses the node (as different users might have different
hold sizes in the future). Currently all hulls have their
hold size set to 36 stacks.