Fix rarity of the ores
The pitchblende is almost nonexistent. After flying around a
world for a few hours and generating over 100 MB of world
data I found just 4 nodes of pitchblende. Gold turned out to
be much more common. Yet the info for these two ores suggests
that the rarity is supposed to be the same.

Additionally, the phosphine clastrate tends to be way too
uncommon.

The problem turns out that the meaning of the "rarity" was
very poorly understood. The reality is that -1.5 may mean
"very rare" for one ore and "almost nonexistent" for
another ore when two ores happen to have it. -2 turns out
to be "nonexistent".

To fix the problem the nitrile ice, carbon oxides ice,
hydrogen sulfide ice and phosphine clathrate were made more
common, creating room for pitchblende to be made less rare
at 1.2. The infotext was also updated to make the range from
"-0.5" to "-1.5" instead of "-1" to "-2".

This also has the nice effect of removing the "challenge" of
having to search for ores you need which are nowhere to be
found. Well, except for the ore market.