Compensate for the much more common pitchblende
To compensate for the heavily reduced rarity of pitchblende
(it is now as common as phosphine clathrate used to be),
increase its weight to whooping 96 tons per cube. This is
explained by some sort of "strange force field influence"
which is responsible also for the ability to have stuff like
stable and maintenance free reactors or small ships with huge
cargo bays.

This makes it much more challenging to dig lots of
pitchblende and bring it to a base for sale. The starting
ship with one pitiful ionic engine will almost be unable to
move.