Fix anchor area flickering when looking around
When the player asks the anchor to show the protected area
and then starts to look around the world, the area sometimes
vanishes and then reappears as he is looking around.

The problem turns out to be a bug in the engine. As the
player is looking around, when the center of the entity is
behind the player, the entity is not displayed even though
there are parts of the entity which are visible.

To fix the problem, 8 entities are created, each having its
"center" in one of the corners of the cube. The code
overlays these entities on top of one another. This bypasses
the bug because either at least one corner is in front of
the player or the entire cube is behind the player.