Fix area display flickering on some locations
When a location of the anchored area has different lighting
levels at its 8 corners, the display of the area flickers
annoyingly when the player looks around.

The problem turns out the 8 entities which compose the
display pick up the differing light levels at their 8
base points (or "center points"), leading to them having
different colors.

This implements a quick and dirty workaround: turn the
color of the frame to black. That prevents the entities from
having different colors, fixing the problem.